+++ src/Exception/StreamException.php
@@ -7,4 +7,28 @@ namespace Solarium\Exception;
  */
 class StreamException extends \UnexpectedValueException implements ExceptionInterface
 {
+    /**
+     * @var string the streaming expression
+     */
+    protected $expression = '';
+
+    /**
+     * Set the streaming expression that caused the exception.
+     *
+     * @param string $expression
+     */
+    public function setExpression(string $expression)
+    {
+        $this->expression = $expression;
+    }
+
+    /**
+     * Get the streaming expression that caused the exception.
+     *
+     * @return string
+     */
+    public function getExpression()
+    {
+        return $this->expression;
+    }
 }